[#11] create mapper, dto

+++ backend/src/main/java/com/Devim/backend/domain/board/BoardType.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class BoardType {
-	private int boardTypeNo;
+	private long boardTypeNo;
 	private String boardName;
 }
